Fixes #51 - removes Guice usage within Jetty server module
