+++ src/main/resources/mesosphere/chaos/rest/impl/logconfig.mustache
