+++ src/main/scala/mesosphere/chaos/metrics/MetricsModule.scala
@@ -3,19 +3,20 @@ package mesosphere.chaos.metrics
 import java.lang.management.ManagementFactory
 
 import com.codahale.metrics.MetricRegistry
-import com.codahale.metrics.jersey.InstrumentedResourceMethodDispatchAdapter
-import com.codahale.metrics.jetty9.InstrumentedHandler
 import com.codahale.metrics.jvm.{ BufferPoolMetricSet, GarbageCollectorMetricSet, MemoryUsageGaugeSet, ThreadStatesGaugeSet }
-import com.google.inject.{ AbstractModule, Provides, Singleton }
-import org.eclipse.jetty.servlet.ServletContextHandler
+import org.slf4j.LoggerFactory
 
-class MetricsModule extends AbstractModule {
-
-  override def configure(): Unit = {}
-
-  @Singleton
-  @Provides
-  def provideRegistry(): MetricRegistry = {
+/**
+  * Includes a registry of metric instances. Those contain:
+  *  - a set of gauges for the counts and elapsed times of garbage collections.
+  *  - a set of gauges for the count, usage, and capacity of the JVM's direct and mapped buffer pools.
+  *  - a set of gauges for JVM memory usage, including stats on heap vs. non-heap memory, plus
+  *    GC-specific memory pools.
+  *  - a set of gauges for the number of threads in their various states and deadlock detection.
+  */
+class MetricsModule {
+  private[this] val log = LoggerFactory.getLogger(getClass.getName)
+  lazy val registry: MetricRegistry = {
     val registry = new MetricRegistry
     registry.register("jvm.gc", new GarbageCollectorMetricSet())
     registry.register("jvm.buffers", new BufferPoolMetricSet(ManagementFactory.getPlatformMBeanServer))
@@ -23,18 +24,4 @@ class MetricsModule extends AbstractModule {
     registry.register("jvm.threads", new ThreadStatesGaugeSet())
     registry
   }
-
-  @Singleton
-  @Provides
-  def provideInstrumentedResourceMethodDispatchAdapter(registry: MetricRegistry) =
-    new InstrumentedResourceMethodDispatchAdapter(registry)
-
-  @Singleton
-  @Provides
-  def provideInstrumentedHandler(servletHandler: ServletContextHandler, registry: MetricRegistry) = {
-    val handler = new InstrumentedHandler(registry)
-    handler.setHandler(servletHandler)
-    handler
-  }
-
 }
